chore: use short primitive instead of int type

+++ implementation/src/main/java/net/velex/title/Title.java
@@ -10,7 +10,7 @@ import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 
 public interface Title {
-  int SERVER_VERSION = serverRelease();
+  short SERVER_VERSION = serverRelease();
   
   /**
    * Returns a new implementation for the {@link net.velex.title.api.AdaptManagerModel} type.
@@ -59,7 +59,7 @@ public interface Title {
    */
   static int serverRelease() {
     String version = Bukkit.getVersion();
-    int index = version.lastIndexOf("MC:");
+    short index = version.lastIndexOf("MC:");
     // Checks if the 'index' value isn't -1.
     // Else, check if the 'version' ends with 'SNAPSHOT'.
     if (index != -1) {
@@ -68,11 +68,11 @@ public interface Title {
       index = version.indexOf('-');
       version = version.substring(0, index);
     }
-    final int lastDot = version.lastIndexOf('.');
+    final short lastDot = version.lastIndexOf('.');
     // Checks if the index of the 'version' value are distinct of 'lastDot' value.
     if (version.indexOf('.') != lastDot) {
       version = version.substring(0, lastDot);
     }
-    return Integer.parseInt(version.substring(2));
+    return Short.parseShort(version.substring(2));
   }
 }
